Make keyword state in parser immutable
Splitting this simpler refactoring out from a larger immutability refactoring: there is no good reason for the keywords used in a given parse context to be mutable. They're used immutably in every real context. As part of this, I've added a reference to System.Collections.Immutable to the language dll, and we will make more use of the immutable collections in future refactorings.