Use an alternate approach to publishing nupkgs only on Windows (#8426)
This avoids an issue where the source build intermediate, built on Linux, didn't have the nupkgs in it because they were only packaged on Windows. Instead of restrictig packaging to only Windows, only push nupkgs on Windows, or in source-build situations. This is a little simpler, and has the added advantage that a developer can work on Linux or Mac and produce the nupkgs for testing or dev work.