Remove add using code action resolver
By allowing our existing code in the formatting pass to run when there are no remappable edits, we can remove the entire add using code action resolver, as we support adding usings from any code action regardless.

I wanted to do this ages ago, but it never worked, and doing this fix I found out why. So here we are.