Add a token for EOF
As we add a new FindToken API to mimic the roslyn rules, we need a token to return when asked for the position at the end of a file. There is no current such token; this change adds an EOF token, which has no content and is cached, to RazorDocumentSyntax.