Remove package refs with PrivateAssets="true" (#8596)
In rzls, the package references, Microsoft.Extensions.Logging and
Microsoft.Extensions.Configuration, were defined with
PrivateAssets="true". (According to git blame, this was done as part
of an attempt to fix spurious build issues.) However, it seems that
this means that the libs in these packages are not included in
published output, which is critical for rzls.exe to run.