Fix issue in CodeWriter when counting line breaks
The CodeWriter counts line breaks as it writes to an underlying
StringBuilder. However, it would count incorrectly if passed a
start index and length representing a portion of a string. Essentially,
it would count past the end of the string portion to the real end of
the string. So, any line breaks appearing after stringIndex + count
would be counted, even though those line breaks wouldn't be written
to the StringBuilder.