Remove StringBuilder from CodeWriter
When building generated code, the Razor compiler regularly passes the
same constant strings multiple times. However, those strings are
 duplicated because `CodeWriter` appends them to a `StringBuilder`,
 which copies the strings into its internal char arrays. Later, the
 strings are duplicated *again* by `StringBuilder.ToString()`.

To avoid the upfront copying and allocation of extra char arrays,
`CodeWriter` now stores strings in a list of `ReadOnlyMemory<char>`
arrays instead of a `StringBuilder`. This ensures that strings are
not copied when `CodeWriter.Write(...)` is called. They are still copied
at the end to produce a string containing all of the generated code.
However, in the future, we will return a more efficient data structure.
https://github.com/dotnet/razor/issues/8076 tracks that work.