Update the UpdateProject method to take a ProjectKey
This is a little problematic, and in fact has the two last uses of ProjectKey.From involved. Right now I could make this call GetAllProjectKeys but I think thats unnecessary, as these places should be able to find the right ProjectKey to use, so I'm leaving updating these until later.