Add reference to Microsoft.CodeAnalysis.CSharp
The language level currently doesn't have a reference to MS.CA.CSharp, which makes many things quite difficult, and resulted in a number of copy/pastes of utility functions from roslyn over to razor. This is the start of addressing that, adding a reference to MS.CA.CSharp and eliminating some of the simpler helpers that were copy/pasted. Contributes to https://github.com/dotnet/razor/issues/8445.