Update various VS bits to use project keys
This essentially adds ProjectKey into how we reason about documents when generating and tracking them in VS, with the exception of the ContainedLanguage project, which is a bit more difficult.

This is the first time I've seen a lot of these components, and whilst I can reason about what they do, I'm not clear on what value they provide, so I'm just blindly updating them.