Add `FindToken` to `SyntaxNode` (#8956)
This implements `SyntaxNode.FindToken`, as was discussed a couple of months ago. The basic semantics of the method are very similar to Roslyn's implementation, but with one significant difference: no attempt is made to ignore any trivia other than whitespace. To recap, this means the rules for `FindToken` are:

1. Find the SyntaxToken that covers the position requested.
2. Scan backwards until a non-whitespace token is encountered. If it's a newline, goto 3. Otherwise, return the token.
3. Scan forwards until a non-whitespace, non-newline token is encountered. Return this token.

There's an additional complication here as well that Roslyn does not have to deal with: because SyntaxToken is a reference type and inherits SyntaxNode, it is possible to call FindToken on that token. If that token is itself a whitespace node, I've taken the opinion that we should not try and do any scanning; we will simply treat it as out of range of the request. I think this is fine for most consumers, as I don't expect there to be any reason for this particular combination of inputs to naturally occur; most calls to FindToken are on the tree root, or on a larger expression. We can adjust this if necessary, however.