Don't mark Utilities.Shared package ref as PrivateAssets="All"
Previously, I had made a change to the ContainedLanguage project to
marrk the Utilities.Shared package ref as PrivateAssets="All".
However, this is problematics for Web Tools unit tests, which will
fail if Utilities.Shared isn't present when exercising
ContainedLanguage. To address this, I'm removing PrivateAssets="All"
and allowing the project reference to flow through as a dependency
of ContainedLanguage. It's not ideal, but this is better than adding
a direct dependency on Utilities.Shared to Web Tools' unit tests.