Improve ProjectRazorJson serialization versioning
ProjectRazorJson had a clumsy approach to versioning its serialization
format. Previously, a "SerializationFormat" property was tacked on at
the very end of the project.razor.json file. However, because it was at
the end of the file, a change to the serialization layout would have
likely caused an exception to be thrown long before the
"SerializationFormat" property was read and checked. In addition, if
the property value was incorrect, the deserializer would just return
null, making it difficult to discover why deserialization failed.

This change adds a "__Version" property to the start of the
project.razor.json file and throws a new ProjectRazorJsonSerializationException
if the property can't be found or its value doesn't match what is
expected. In addition, ProjectRazorJson serialization has been updated
to expect properties in a particular order, rather than allowing the
order to be loose.