Remove some uses of the dispatcher (#9023)
Now that the project snapshot manager is a lock based management, reading especially should be easy to remove. This removes it from two areas to start, but it should be fairly impactful.

The DocumentContextFactory gets used on every LSP incoming message to create the context we use for our handlers. Hopefully we should just see wins for this removal, since we're now consistant with expectations around our message dispatching. Things that do not mutate the workspace are allowed to run in parallel, and things that don't should be run serial and block other handlers from getting called.

The removal from the component search seems trivial, so I just did it. I'm not sure of the impact of this area tbh, but it shouldn't break anything.