Fix OSSkipConditionFactAttribute to work correctly on .NET Framework
The way our OSSkipConditionFactAttribute was implemented does work
for "Windows" when running on .NET Framework. This is due to the way
that `RuntimeInformation.IsOSPlatform(...)` is written for .NET
Framework. Instead of comparing strings, it performs a reference
comparison for the give `OSPlatform` to `OSPlatform.Windows`. Since
the code creates new `OSPlatform` with the string "Windows" instead of
using `OSPlatform.Windows`, it fails.