Remove a couple duplicated GetContent calls (#9123)
ComponentFileKindVisitor.VisitMarkupAttributeBlock's call to GetContent is duplicated and appears to be about 0.8% of allocations in the customer profile I'm looking at. I looked at the IL in a release version of the assembly and the compiler at least doesn't remove the call (unclear whether the jitter would be able to). Either way, it's clearer and potentially faster to remove the duplicated calls.

Didn't look into whether any of the remaining 1.2% of allocations due to calling GetContent could be optimized.