Use dispatcher for project snapshot change notifications (#9165)
Fix for AB#1866282

It's possible for the SnapshotResolver to create the MiscellaneousProject when trying to resolve a snapshot in any project. While that's nice, and intended, some code still depends on the dispatcher thread when receiving the notification for change. This change makes sure change calls happen on the dispatcher thread, but read/write are still behind locks.

Filed #9162 to actually fix this, but this fix should work for P2 while we make more complex fixes for P3