Ensure that directive nodes always have at least one token child (#9195)
* Ensure that directive nodes always have at least one token child

Fixes https://github.com/dotnet/razor/issues/9177. The parser was violating an invariant that nodes cannot be terminal nodes, and must always have at least one token child. This is likely not the only variant of this problem, but is a very common one, as typing in the editor will often result in incomplete directives some of the time. We now add a missing token in such cases to avoid violating the invariant.

* Update IncompleteDirectives test

* More baseline updates