Simplify assembly filtering in tag helper collector
The check to determine whether an assembly should be considered in
tag helper and component discovery was incorrect. First, it did not
consider transitive assembly references, and second, it only considered
the assembly name. It turns out that getting this check to be correct
is challenging. So, I'm switching this back to what it was before the
assembly caching change; it just checks whether the assembly name
starts with "System.".