Fix RemoteInvocationException reporting (#9596)
The initial IFaultExceptionHandler MEF construction was no longer working due to refactoring. This changes it to no longer use MEF and just report directly.

Also update the ExceptionStrategy to ISerializable, since that will set the InnerException property if the other end of the connection is the same. This will make error reporting for us contain remote stacks instead of our own. If that's not the case, we still report as much as we can.

Anyone looking for more information on jsonrpc errors can look at microsoft/vs-streamjsonrpc@main/doc/exceptions.md