Refresh settings when unified settings change (#9737)
Finalizes AB#1786418

Unified settins has been in for a while, but we needed to consume the 17.9 preview packages to be able to subscribe to the change events. This adds that, updates the structures we use to help keep it easy to follow, and adds a readme for people adding new options.