Unify IErrorReporter instances
This change unifies the IErrorReporter instances used in Razor tooling.
It turns out that there were extras that might cause some errors to
not be reported at all.

1. I've removed the `ErrorReporter.Instance` property. This was a
simple `IErrorReporter` instance that did nothing at all. However, it
was passed into the `ProjectSnapshotManager` used in the language
server. So, any errors from the server's `ProjectSnapshotMangager`
wouildn't actually be reported anywhere useful. Now, the server's
`ProjectSnapshotManager` will use the correct `IErrorReporter` instance,
which reports errors to an `ILogger`.
2. `IErrorReporter` is no longer a Roslyn workspace service. So,
everyplace that retrieved an instance from the workspace is updated to
import an `IErrorReporter` from the DI host (i.e. MEF in VS, and
MS.Extensions.DependencyInjection in the language server).
3. In VS, because `IErrorReporter` isn't exported as a Roslyn workspace
service, that means that there are no longer _two_ `IErrorReporter`
instances in VS. There's now just the one.