Add telemetry for time-to-first diagnostic (#9795)
* Add telemetry for time-to-first diagnostic

By adding this Diagnostic kind to our set of capabilities, the VS Win system lights up to gather the time to first diagnostic event, so we can track changes in that time.  This will effectively be for our entire diagnostic process, not just syntactic diagnostics as the name of the constants implies.
Adds a simple unittest for the endpoint.