Don't use Workspace in WindowsRazorProjectHostBase
This change avoids passing a Roslyn workspace into
`WindowRazorProjectHostBase`. This was only used to retrieve the
`ProjectSnapshotManager`. However, `IProjectSnapshotManagerAccessor`
exists for that purpose.

Note that I've had to call `JTF.Run(...)` in the VS implementation
of `IProjectSnapshotManagerAccessor.Instance` to avoid the
snapshot manager being created and initializing project snapshot
change triggers on a non-dispatcher thread. I made this change after
calling `AssertDispatcherThread()` in `IProjectSnapshotManagerAccessor`
and finding that it was indeed being initialized from a different thread
(the UI thread) by `CSharpVirtualDocumentFactory`. Because
`WindowsRazorProjectHostBase` had a comment about only creating the
snapshot manager on the dispatcher thread, I decided to fix this
violation of that rule with a `JTF.Run(...)`. I do intend to completely
refactor the way that change triggers work, so this is expected to be
a temporary change for correctness.