Fixes #9812 and #9906
ISSUE:
Because we are are keying off of the intermediate object path instead of the csproj file path, the intermediate object path may have ".." in it from normal MSBuiild properties, but that means the same logical string has two different representations.

FIX:
Normalize the path.  There's an adjoining and similiar change in Roslyn to likewise make.  (Will link).  There is alternative fix to pursue where Razor instead does the  normalization.

TESTING:
Unittests pulled from manul testing with ".." in BaseIntermediateOutputPath and ArtifactsPath