Dispatcher should run actions inline if possible
This is a pretty simple change but should reduce deadlocks and improve
the performance of running operations on the dispatcher. Essentially,
if the current execution path is already running on the dispatcher, new
operations can be run inline without starting a task.

In addition, I'm cleaned up and rationalized the `RunAsync(...)`
methods a bit.