Tweak RazorConfigurationFormatter to account for serialization format change
The recent FUSE changes updated the RazorConfigurationFormatter to write serialize/deserialize a new boolean. However, `SerializationFormat.Version` for MessagePack wasn't updated, so it just fails during deserialization when older project.razor.bin files are initially read off disk. This tweaks the deserializer to be more forgiving and handle the case that the boolean isn't present.

A similar change is not needed in the JSON serialization because that deserializer was written to be a bit more forgiving and the JSON `SerializationFormat.Version` was updated.