Remove lock in BackgroundDocumentGenerator
There is a fair amount of lock contention in `BackgroundDocumentGenerator` that contributes to a hang on 17.10p2. The issue is caused by my recent changes to `FallbackProjectManager` to perform `IProjectSnapshotManager` updates on the dispatcher. While this change improves correctness it changes the interaction with Roslyn's dynamic file info implementation enough that it is causing problems.

The particular issue that this should address locks on `BackgroundDocumentGenerator._suppressedDocuments`.