Don't get IVsAsyncFileChangeEx on background thread
`VisualStudioFileChangeTrackerFactory` retrieves an `IVsAsyncFileChangeEx` service in its constructor. Because this is an `[ImportingConstructor]` it will likely be called on the thread pool and implicitly marshal to the UI thread to perform a QI as part of retrieving the service. This can cause a hang if the UI thread is tied up with other work, such as when a solution is loaded.

This change avoids retrieving `IVsAsyncFileChangeEx` until its needed. This does block on retrieving the service. However, avoiding that will require a more substantial change that I'll follow up with on `main`.