Fix vscode fuse feature flag (#10169)
Prior to this change, the feature flag was being serialized and deserialized as part of the RazorConfiguration. This works fine for VS, since the razor extension is the one serializing. However, if we want to store that information in the serialization on vs code, we'd have to have a hook into Roslyn to do that.

Instead, this adds a section of the configuration specifically labeled for feature flags. This should not be serialized, and should be filled in based on the LanguageServerConfiguration. This allows us to pass the flag into rzls and retrieve for the correct implementation when running the ProjectEngine