Avoid fault where misc files project is added off dispatcher
Telemetry recently indicated an assert is triggering in certain
scenarios if the snapshot resolver adds the misc files project off
of the dispatcher. To avoid this, I've updated the snapshot resolver
to immediately start adding the misc files project when its created.
Then, when the misc files project is requested, we'll wait for it to be
added if it hasn't been yet. This is a temporary solution that will go
away once project management is made more async.