Ensure that RazorProjectService operations happen serially
Fixes #10238

Bad things can happen if a project is added while a document is opened,
so we shouldn't allow it. Likewise, other RazorProjectService operations
shouldn't run at the same time. This is achieved with an
`AsyncSemaphore`, which is not re-entrant. So, it's important that the
RazorProjectService operations do not rely on other RazorProjectService
operations directly.