Don't use batching work queue in ProjectConfigurationStateSynchronizer
Fixes #10306

The recent change to use an `AsyncBatchingWorkQueue` in
`ProjectConfigurationStateSynchronizer` has resulted in a signficant
regression. The syncrhonizer has somewhat subtle semantics that doesn't
mesh well with `AsyncBatchingWorkQueue`. As part of rolling this change
back, I've commented the code a bit more thoroughly to call out how it
works and where the semantics are a bit flaky. Ultimately, I expect
we'll be able to get rid of this once we stop writing
`project.razor.vs.bin` files to disk.