Re-think how Razor options get loaded in Visual Studio
This represents a significant change to how options are loaded in VS. Rather than exporting  `LanguageServerFeatureOptions` directly, a new `ILanguageServerFeatureOptionsProvider` abstraction has been introduced to provide an indirection for acquiring options. The Visual Studio implementation of this interface uses a VS `IServiceProvider` to request a new `IVisualStudioOptionsProvider` service that is proferred by the Razor package. This service does the work of acquiring `IVsFeatureFlags` and `ISettingsManager`.