Make it easy to include NetCore-only and NetFx-only files
This migrates a change from Razor tooling and makes it available for all the Razor repo. By naming a C# file as `.NetCore.cs` or `.NetFx.cs` it will compile _only_ for NetCore or NetFx, respectively. Alternatively, files can be placed in a folder called `NetCore` or `NetFx`, or with a name that ends in `_NetCore` or `_NetFx` to control the target framework they're compiled for. This makes it clearer from the Solution Explorer if there are files that only compile for one target or another rather than having to look at `#if..#endif` blocks.