Pool CodeWriter ReadOnlyMemory<char> pages
These allocations are present in a customer trace I'm looking at, accounting for 1.4% of allocations in the VS process (around 100 MB).

The owner of the CodeWriter is already disposable, so making the CodeWriter disposable is trivial, and allows for all ReadOnlyMemory<char> pages added to _pages to be released back to a pool.