Removing unnecessary Razor Legacy Editor feature flag
AFAIK it was introduced for new LSP editor rollout several years ago before we had Tools/Options/HTML Editor/Advanced/Use legacy Razor Editor option. Now that rollout is over, and there is a public non-preview mechanism for users to switch to the old editor if needed, the feature flag is no longer necessary and creates confusion (and extra code to test/maintain).