Add latest .NET 8 runtime to global.json
Since we updated to a .NET 9 SDK, the
MS.ANC.Razor.Microbenchmarks.Generator unit tests have failed to run
locally. This is because these test use Benchmark.NET, which performs a
builder at runtime targeting .NET 8. This works fine in CI because it
uses a .NET 8 SDK.

This change just adds the most recent .NET 8 runtime version to the
list of runtimes in Razor's global.json. So, it'll get installed into
the .dotnet folder and the Microbenchmarks.Generator will run and stop
reporting a spurious error.