Replace IProjectCollectionResolver with IProjectQueryService
IProjectCollectionResolver was intended to be passed into services that need to enumerate project snapshots to allow them to be used in co-hosting. Instead of encapsulating the state it operates on, a "context" IDocumentSnapshot was passed into IProjectCollectionResolver.EnumerateProjects(...). In co-hosting, the "context" IDocumentSnapshot would be used to extract the Roslyn Solution, while the language server would ignore the parameter entirely.

This change introduces IProjectQueryService to replace IProjectCollectionResolver. In the language server, it wraps IProjectSnapshotManager. In co-hosting, a new instance of the service is created for any service call and wraps the Roslyn Solution snapshot.