Stop trying to guess which projects documents belong to
When publishing, we see a project remove for obj/Debug, and move all files to the misc
project, then we'd see a new project for obj/Release, and add that. That add would then
cause us to go through all of the misc files, and add them to the first project that could
possibly have them - in this case, adding them right back to obj/Debug.

Generally speaking we shouldn't be guessing at anything with project info, but rather rely
on correct input from systems. ie, if they add a project, they presumably will then follow
up with telling us which documents are in it.