Update AssumeNotNull extensions to include optional message
Because this change modifies the AssumeNotNulll extension methods to defer to Assumed.NonNull, it's necessary to update TelemetryReporter to include the ThrowHelper and Assumed classes when performing its stack walk for faults.

I suspect that these types should have already been included anyway, to improve the fault bucketing. It might be useful to add ArgHelper as well, but that's not really in scope here.